Add notification on failed save
Since saving the Tiddlywiki is very important to the user and losing the
progress made may be very impactful, the addon should tell the user if
any error has been detected.  This commit adds the "notification"
permission to manifest.json and displays a browser notification on any
error detected while sending a native message to the native application.

Since the user may save the Tiddlywiki file often, no notification is
shown in the event of a successful save: this is by design to avoid
training the user to ignore the notification.